[fix] 테마 스위치: 멘토님 피드백 반영
