+++ src/Router.js
@@ -1,8 +1,5 @@
-import React, { useState } from 'react';
-// import { ThemeProvider } from 'styled-components';
-// import { lightTheme, darkTheme } from './styles/theme';
-// import ThemeSwitcher from './modules/themeSwitcher';
-// import GlobalStyle from './styles/GlobalStyle';
+import React from 'react';
+import App from './App';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Main from './pages/Main/Main';
 import Login from './pages/Login/Login';
@@ -10,6 +7,7 @@ import Login from './pages/Login/Login';
 const Router = () => {
   return (
     <BrowserRouter>
+      <App />
       <Routes>
         <Route path="/" element={<Main />} />
         <Route path="/login" element={<Login />} />
