+++ src/pages/Main/Main.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const Main = () => {
-  return <>1</>;
+  return <>매인</>;
 };
 
 export default Main;
